Fixed issue with file not found issues

+++ src/AutoTest.Core/BuildRunners/MSTestCrossPlatformPreProcessor.cs
@@ -29,19 +29,30 @@ namespace AutoTest.Core.BuildRunners
                                  }
                                  else
                                  {
-                                     var project = File.ReadAllText(x.Project.Key);
-                                    if (switcher.IsGuyInCloset(project))
-                                    {
-                                        var tmpProject = getTempProject(x.Project.Key);
-                                        // We can do this because we know they are ordered in the right order
-                                        _tmpProjects.ForEach(y => 
-                                            project = project
-                                                .Replace("\\" + Path.GetFileName(getOriginalProject(y)), "\\" + Path.GetFileName(y))
-                                                .Replace("\"" + Path.GetFileName(getOriginalProject(y)), "\"" + Path.GetFileName(y)));
-                                        File.WriteAllText(tmpProject, switcher.PerformSwitch(project));
-                                        _tmpProjects.Add(tmpProject);
-                                        x.BuildTemporaryProject(tmpProject);
-                                    }
+                                     if (File.Exists(x.Project.Key))
+                                     {
+                                         var project = File.ReadAllText(x.Project.Key);
+                                         if (switcher.IsGuyInCloset(project))
+                                         {
+                                             var tmpProject = getTempProject(x.Project.Key);
+                                             // We can do this because we know they are ordered in the right order
+                                             _tmpProjects.ForEach(y =>
+                                                                  project = project
+                                                                                .Replace(
+                                                                                    "\\" +
+                                                                                    Path.GetFileName(
+                                                                                        getOriginalProject(y)),
+                                                                                    "\\" + Path.GetFileName(y))
+                                                                                .Replace(
+                                                                                    "\"" +
+                                                                                    Path.GetFileName(
+                                                                                        getOriginalProject(y)),
+                                                                                    "\"" + Path.GetFileName(y)));
+                                             File.WriteAllText(tmpProject, switcher.PerformSwitch(project));
+                                             _tmpProjects.Add(tmpProject);
+                                             x.BuildTemporaryProject(tmpProject);
+                                         }
+                                     }
                                  }
                              });
             return details;
